fix(dashboard): fix select bug and refactor graphs
