+++ src/components/Dashboard/Charts/index.tsx
@@ -1,0 +1,3 @@
+export { default as ClosedSupportChart } from "./ClosedSupportChart";
+export { default as StatusUpdateChart } from "./StatusUpdateChart";
+export { default as GeneralComparisonChart } from "./GeneralComparisonChart";
