--- src/components/Dashboard/Graphics/index.tsx
@@ -1,3 +1,0 @@
-export { default as ClosedSupportGraphic } from "./ClosedSupportGraphic";
-export { default as StatusUpdatesGraphic } from "./StatusUpdatesGraphic";
-export { default as GeneralComparisonGraphic } from "./GeneralComparisonGraphic";
