--- src/utils/analyticsStyles.ts
@@ -1,4 +1,0 @@
-export const analyticsColors = {
-  primary: '#8787bb',
-  secondary: '#f0d4d4',
-};
