+++ src/utils/definitions.ts
@@ -109,8 +109,8 @@ export type ChatContextShared = {
   setConversation: Dispatch<SetStateAction<Conversation>>;
 };
 
-export type PeriodData = {
-  closedSupportGraphic: {
+export type FilteredAnalyticsData = {
+  closedSupportChart: {
     total: {
       porcentage: number;
       value: number;
@@ -120,25 +120,23 @@ export type PeriodData = {
       value: number;
     };
   };
-  generalComparisonGraphic: {
+  generalComparisonChart: {
     general: number[];
     own: (number | null)[];
   };
-  statusUpdatesGraphic: {
+  statusUpdateChart: {
     openSupport: number[];
     closedSupport: number[];
   };
 };
 
 export type AnalyticsData = {
-  yesterday: PeriodData;
-  last_30_days: PeriodData;
-  last_6_months: PeriodData;
-  last_12_months: PeriodData;
+  [key: string]: FilteredAnalyticsData;
 };
 
 export type AnalyticsContextShared = {
   analytics: AnalyticsData;
-  filteredData: PeriodData;
+  colors: { [key: string]: string };
+  filteredData: FilteredAnalyticsData;
   setFilter: Dispatch<SetStateAction<string>>;
 };
